ingest regions
