+++ apps/f3-glossary/package.json
@@ -50,6 +50,7 @@
     "@radix-ui/react-toggle": "1.1.1",
     "@radix-ui/react-toggle-group": "1.1.1",
     "@radix-ui/react-tooltip": "1.1.6",
+    "@types/lodash": "^4.17.17",
     "@types/pg": "^8.15.2",
     "autoprefixer": "^10.4.20",
     "class-variance-authority": "^0.7.1",
@@ -60,6 +61,7 @@
     "drizzle-orm": "^0.44.1",
     "embla-carousel-react": "8.5.1",
     "input-otp": "1.4.1",
+    "lodash": "^4.17.21",
     "lucide-react": "^0.454.0",
     "next": "15.2.4",
     "next-themes": "^0.4.4",
